move src-loader to utils/texture-loader
